requirements.txt
Django==1.11.6
djangorestframework==3.7.0
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
pytz==2017.2
whitenoise==3.3.1
